refactor utils lib

+++ cpme_api/api/feature/utils.py
@@ -1,12 +1,8 @@
-import time
-import enum
 import os
-import functools
 from pprintjson import pprintjson
 import json
 from json.decoder import JSONDecodeError
-from model_gen.config import settings, Environment
-from typing import Tuple, Any, List, TextIO
+from typing import List, TextIO
 import logging
 import csv
 
@@ -14,72 +10,12 @@ import csv
 log = logging.getLogger('cpme_api')
 
 
-class Timer(object):
-    def __init__(self):
-        self.main_timer = None
-        self.timer = None
-
-    def start(self):
-        self.main_timer = time.time()
-        self.timer = time.time()
-
-    def reset(self, main_timer=False):
-        if main_timer:
-            self.main_timer = time.time()
-        else:
-            self.timer = time.time()
-
-    def get_time_str(self, main_timer=False):
-        end = time.time()
-        if main_timer:
-            if not self.main_timer:
-                self.main_timer = end
-            start_time = self.main_timer
-        else:
-            if not self.timer:
-                self.timer = end
-            start_time = self.timer
-        m, s = divmod(end - start_time, 60)
-        h, m = divmod(m, 60)
-        return "%02dh:%02dm:%02ds" % (h, m, s)
-
-
-def data_folder_fix(path: str):
-    """Check if a folder variable name is relative
-    path starts with . and add global SCENARIO_PATH_FOLDER."""
-    if path.startswith('./'):
-        path = os.path.join(settings.scenario_path_folder, path.lstrip("./"))
-    return path
-
-
-class UnitSize(enum.Enum):
-    BYTES = 1
-    KB = 2
-    MB = 3
-    GB = 4
-
-
-def convert_unit(size_in_bytes: int, unit=UnitSize.BYTES):
-    """ Convert the size from bytes to other units like KB, MB or GB"""
-    if unit == UnitSize.KB:
-        return size_in_bytes/1024
-    elif unit == UnitSize.MB:
-        return size_in_bytes/(1024*1024)
-    elif unit == UnitSize.GB:
-        return size_in_bytes/(1024*1024*1024)
-    else:
-        return size_in_bytes
-
-
-def get_file_size(file_path: str, size_type=UnitSize.MB):
-    """ Get file in size in given unit like KB, MB or GB"""
-    assert os.path.isfile(file_path), f'{file_path} not found'
-    size = os.path.getsize(file_path)
-    return round(convert_unit(size, size_type), 3)
-
-
 def fancy(json_dict: dict, line_text: str = ''):
-    """Pretty print of dict structure into stdout only!"""
+    """
+    Pretty print of dict structure as a JSON format into stdout only!
+    :param json_dict: Dictionary data structure.
+    :param line_text: A text to print before data structure.
+    """
     if json_dict is None:
         print('\tNot defined!\n')
         return
@@ -93,42 +29,75 @@ def fancy(json_dict: dict, line_text: str = ''):
         print('\tempty {}\n')
 
 
-def body_request_file_name(ep: str,
-                           component_type: str,
-                           env: str,
-                           prefix: str = '',
-                           suffix='json') -> str:
-    if prefix:
-        prefix += '_'
-    return f'{prefix}{ep}.{suffix}'
+def dir_check(file_dir: str, del_files: bool = False):
+    if not file_dir:
+        return
+
+    if not os.path.isdir(file_dir):
+        log.warning(f'Creating the new directory {os.path.abspath(file_dir)}')
+        os.makedirs(file_dir, exist_ok=True)
+
+    if del_files:
+        for file in os.listdir(file_dir):
+            if os.path.isfile(file_dir + file):
+                os.remove(file_dir + file)
 
 
-def json_to_file(data_dict: dict, file_name: str, verbose=True, update=True):
+def json_to_file(data_dict: dict, file_name: str, verbose=True, update=True) -> str:
+    """
+    Save dictionary structure into the JSON file.
+    :param data_dict: dictionary object of json response
+    :param file_name: The file_name can be string with absolute ('/.') or relative ('./') path or just a file name.
+    :param verbose:  if False disable information messages
+    :param update: if True it overwrites already existing file
+    :return: absolute path of file_name
+    """
+    dir_name = os.path.dirname(file_name)
+    dir_exist = os.path.isdir(dir_name)
+
     if not file_name.endswith('.json'):
         file_name = file_name + '.json'
-    if update:
+
+    if dir_name == '':
+        dir_exist = True
+        file_name = './example/' + file_name
+
+    file_name = os.path.abspath(file_name)
+    if update and dir_exist:
         with open(file_name, 'w') as f:
             json.dump(data_dict, f, indent=4, sort_keys=True)
-        if verbose:
-            log.info(f'Response saved in {file_name}')
-    elif path := find_file(path=file_name):
+    elif path := find_file(path=file_name, folder=dir_name):
         if path:
-            log.info(f'Response already stored {path}')
+            log.info(f'The file name already exists {path}. Use update=True to overwrite it.')
+        return file_name
     else:
+        dir_check(dir_name)
         with open(file_name, 'w') as f:
             json.dump(data_dict, f, indent=4, sort_keys=True)
-        if verbose:
-            log.info(f'Response saved in {file_name}')
+
+    if verbose:
+        log.info(f'The content saved into {file_name}')
     return file_name
 
 
 def load_json_file(file_dir: str = '', file_name: str = '', path: str = None, verbose: bool = False) -> dict:
+    """
+
+    :param file_dir:
+    :param file_name:
+    :param path:
+    :param verbose:
+    :return:
+    """
     """Load JSON data into dict."""
     if path is None:
         path = os.path.join(file_dir, file_name)
+    path = os.path.abspath(path)
+
     if not os.path.isfile(path):
         log.error(f"Not such file or directory {path}!")
         return {}
+
     try:
         with open(path) as f:
             if verbose:
@@ -139,98 +108,26 @@ def load_json_file(file_dir: str = '', file_name: str = '', path: str = None, ve
         return {}
 
 
-def dir_check(file_dir: str, del_files: bool = False):
-    if not os.path.isdir(file_dir):
-        log.warning(f'Creating the new directory {file_dir}')
-        os.makedirs(file_dir, exist_ok=True)
-    if del_files:
-        for file in os.listdir(file_dir):
-            if os.path.isfile(file_dir + file):
-                os.remove(file_dir + file)
-
-
-def align_prefix(prefix: str = ''):
-    if prefix:
-        return prefix + '_'
-    else:
-        return ''
-
-
-def update_prefix_content(prefix: str) -> str:
-    if '_' in prefix:
-        prefix = prefix.replace('_', '-')
-    return prefix
-
-
-def files_in_folder(folder=data_folder_fix(settings.requests_folder)):
-    filenames = []
-    for (_, _, filenames) in os.walk(folder):
-        break
-    return sorted(filenames)
-
-
-def find_file_by_prefix(prefix: str, folder=data_folder_fix(settings.requests_folder)):
-    prefix = update_prefix_content(prefix)
-    f_files = []
-    for file in files_in_folder(folder):
-        if prefix in file:
-            f_files.append(file)
-    if len(f_files) == 1:
-        return f_files[0]
-    elif (l:=len(f_files)) > 1:
-        log.warning(f'Found {l} files {f_files}. Loaded only the first one!')
-        return f_files[0]
-    else:
-        return ''
-
-
-def find_file(prefix: str = None,
-              folder: str = '',
+def find_file(folder: str = '',
               file_name: str = '',
               path: str = '') -> str:
-    if prefix:
-        prefix = update_prefix_content(prefix)
     if not folder:
-        folder = data_folder_fix("./example/data/")
-    else:
-        folder = data_folder_fix(folder)
+        folder = "./example/data/"
     if os.path.isfile(path):
         pass
     elif os.path.isfile(os.path.join(folder, file_name)):
         path = os.path.join(folder, file_name)
-    elif prefix and (file_name := find_file_by_prefix(prefix, folder)):
-        path = os.path.join(folder, file_name)
     else:
-        if prefix:
-            log.warning(f'File with prefix = {prefix} not found in {folder}!')
-        elif path:
+        if path:
             log.warning(f'File {path} not found!')
         else:
             log.warning(f'File {file_name} not found in {folder}!')
         return ''
-    return path
-
-
-def data_from_path(path: str) -> Tuple[str, str]:
-    f_format = ''
-    if not os.path.isfile(path):
-        return None
-    if '.json' in path:
-        data = load_json_file(path=path)
-        f_format = 'json'
-    elif '.csv' in path:
-        data = csv_to_str(path)
-        f_format = 'csv'
-    elif '.xml' in path:
-        data = xml_to_str(path)
-        f_format = 'xml'
-    else:
-        log.warning('Not supported file format [csv,xml,json]')
-        data = ''
-    return data, f_format
+    return os.path.abspath(path)
 
 
 def xml_to_str(path: str) -> str:
+    """return content of xml file"""
     if not path.endswith('.xml'):
         return ''
     with open(path) as f:
@@ -238,152 +135,20 @@ def xml_to_str(path: str) -> str:
 
 
 def csv_to_str(path: str) -> str:
+    """return content of csv file"""
     if not path.endswith('.csv'):
         return ''
     with open(path) as f:
         return f.read()
 
 
-def truncate_list(vector: list, max_size: int = 100000) -> list:
-    """Truncate list size with max_size"""
-    assert isinstance(vector, list), 'Vector must by instance of list!'
-    if len(vector) < max_size:
-        max_size = len(vector)
-    return vector[:max_size]
-
-
-def isfloat(num: str) -> bool:
-    try:
-        float(num)
-        return True
-    except ValueError:
-        return False
-
-
-def is_float(value) -> bool:
-    if isinstance(value, float):
-        return True
-    else:
-        return False
-
-
-def isnumber(value: str):
-    if value is None:
-        return value
-    if isinstance(value, int) or isinstance(value, float):
-        return value
-    if 'False' == value:
-        return False
-    elif 'True' == value:
-        return True
-    if value.isdigit():
-        return int(value)
-    elif isfloat(value):
-        return float(value)
-    else:
-        return value
-
-
-def catch_exception(enable: bool = False) -> Tuple[Any, bool]:
-    """Catch Comet exceptions if enable == True. And return tuple of result state"""
-    def decorator(func):
-        if not enable:
-            return func
-
-        @functools.wraps(func)
-        def raiser(*args, **kwargs):
-            try:
-                res = func(*args, **kwargs)
-                if res or res is None:
-                    print(f'func {func.__name__}: result => PASSED')
-                else:
-                    print(f'func {func.__name__}: result => FAILED')
-                return res, True
-            except Exception as e:
-                print(f'func {func.__name__}: result => FAILED EXP_MSG:{e}')
-                return e, False
-        return raiser
-
-    return decorator
-
-
-def env_equal(envs: list = []):
-    al_env = [Environment.alias(e) for e in envs]
-    if envs is [] or (settings.current_env in al_env):
-        return True
-    else:
-        return False
-
-
-def use_for_env(envs: list = []):
-    """Catch Comet exceptions if enable == True. And return tuple of result state"""
-    def decorator(func):
-        al_env = [Environment.alias(e) for e in envs]
-        if envs is [] or (settings.current_env in al_env):
-            return func
-
-        @functools.wraps(func)
-        def raiser(*args, **kwargs):
-            log.warning(f'Skipping testcase {func.__name__} because it is valid only for {al_env}')
-            return
-            #res = func(*args, **kwargs)
-
-        return raiser
-
-    return decorator
-
-
-def narrative(func):
-    """decorator to print a doc string"""
-    @functools.wraps(func)
-    def raiser(*args, **kwargs):
-        log.info(f'[ {func.__name__} ]: ===SCENARIO DESCRIPTION=== < {func.__doc__} >')
-        return func(*args, **kwargs)
-
-    return raiser
-
-
-def check_response_model(func):
-    def pattern(*args, **kwargs):
-        class Status:
-            def __init__(self):
-                self._attrs = list()
-                self._register = dict()
-
-            def exist(self, model):
-                self._attrs.append(model.id)
-
-            def add(self, model):
-                self._register[model.id] = str(model)
-
-        status = Status()
-        kwargs.update({'state': status})
-        func(*args, **kwargs)
-        res = list()
-        tc_num = int()
-        for key in status._register:
-            tc_num += 1
-            if key in status._attrs:
-                log.info(f'TestSuite-{func.__name__}<TestCase-{tc_num}>: {key} found, records={status._attrs.count(key)}')
-            else:
-                res.append(f'Attribute Not found {status._register[key]}')
-                log.error(f'TestSuite-{func.__name__}<TestCase-{tc_num}>: '
-                          f'Attribute Not found {status._register[key]}')
-        if res:
-            raise Exception('\n'.join(res))
-        return True
-    return pattern
-
-
 def load_csv_to_list(path: str = '',
                      folder: str = '',
-                     file_name: str = '',
-                     prefix: str = None) -> List[dict]:
+                     file_name: str = '') -> List[dict]:
     """"Load list of dictionaries from csv file."""
     path = find_file(file_name=file_name,
                      path=path,
-                     folder=folder,
-                     prefix=prefix)
+                     folder=folder)
     if path == '':
         log.warning(f"File not found")
         return []
